Fix realized metadata for shadowed capabilities
