+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/external/model/AbstractRealisedModuleResolveMetadataSerializationHelper.java
@@ -137,7 +137,12 @@ public abstract class AbstractRealisedModuleResolveMetadataSerializationHelper {
         int capabilitiesCount = decoder.readSmallInt();
         List<Capability> rawCapabilities = Lists.newArrayListWithCapacity(capabilitiesCount);
         for (int j = 0; j < capabilitiesCount; j++) {
-            rawCapabilities.add(new ImmutableCapability(decoder.readString(), decoder.readString(), decoder.readString()));
+            String appendix = decoder.readNullableString();
+            CapabilityInternal capability = new ImmutableCapability(decoder.readString(), decoder.readString(), decoder.readString());
+            if (appendix != null) {
+                capability = new DefaultShadowedCapability(capability, appendix);
+            }
+            rawCapabilities.add(capability);
         }
         return ImmutableCapabilities.of(rawCapabilities);
     }
@@ -147,6 +152,14 @@ public abstract class AbstractRealisedModuleResolveMetadataSerializationHelper {
     private void writeCapabilities(Encoder encoder, List<? extends Capability> capabilities) throws IOException {
         encoder.writeSmallInt(capabilities.size());
         for (Capability capability: capabilities) {
+            boolean shadowed = capability instanceof ShadowedCapability;
+            if (shadowed) {
+                ShadowedCapability shadowedCapability = (ShadowedCapability) capability;
+                encoder.writeNullableString(shadowedCapability.getAppendix());
+                capability = shadowedCapability.getShadowedCapability();
+            } else {
+                encoder.writeNullableString(null);
+            }
             encoder.writeString(capability.getGroup());
             encoder.writeString(capability.getName());
             encoder.writeString(capability.getVersion());
